refactor: generator 패키지로 분리
