+++ src/main/java/ladder/domain/generator/LadderStepGenerator.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.generator;
 
 import java.util.List;
