+++ src/test/java/ladder/domain/generator/LadderStepGeneratorTest.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.generator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
